+++ SpringWebApp/src/main/java/isamrs/controller/DijagnozaController.java
@@ -3,6 +3,8 @@ package isamrs.controller;
 import java.util.Collection;
 import java.util.stream.Collectors;
 
+import isamrs.domain.AdministratorKlinickogCentra;
+import isamrs.domain.Lekar;
 import isamrs.dto.DijagnozaDTO;
 import isamrs.exceptions.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,6 +24,8 @@ import org.springframework.web.server.ResponseStatusException;
 import isamrs.domain.Dijagnoza;
 import isamrs.service.DijagnozaService;
 
+import javax.servlet.http.HttpServletRequest;
+
 @RestController
 @RequestMapping("/dijagnoza")
 public class DijagnozaController {
@@ -29,14 +33,20 @@ public class DijagnozaController {
 	private DijagnozaService service;
 	
 	@GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<Collection<DijagnozaDTO>> getDijagnoze(){
+	public ResponseEntity<Collection<DijagnozaDTO>> getDijagnoze(HttpServletRequest request){
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra)&&!(request.getSession().getAttribute("user") instanceof Lekar)){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		Collection<DijagnozaDTO> dijagnoze = service.findAll().stream().map(DijagnozaDTO::new).collect(Collectors.toList());
 		
 		return new ResponseEntity<Collection<DijagnozaDTO>>(dijagnoze, HttpStatus.OK);
 	}
 	
 	@GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<DijagnozaDTO> getDijagnoza(@PathVariable("id") Long id){
+	public ResponseEntity<DijagnozaDTO> getDijagnoza(@PathVariable("id") Long id, HttpServletRequest request){
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra)){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		try {			
 			Dijagnoza d = service.findOne(id);
 			return new ResponseEntity<DijagnozaDTO>(new DijagnozaDTO(d), HttpStatus.OK);
@@ -47,19 +57,28 @@ public class DijagnozaController {
 	}
 	
 	@PostMapping(produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<DijagnozaDTO> createDijagnoza(@RequestBody Dijagnoza d){
-		Dijagnoza d2 = service.create(d);
+	public ResponseEntity<DijagnozaDTO> createDijagnoza(@RequestBody DijagnozaDTO d, HttpServletRequest request){
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra)){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		Dijagnoza d2 = service.create(new Dijagnoza(d));
 		return new ResponseEntity<DijagnozaDTO>(new DijagnozaDTO(d2), HttpStatus.CREATED);
 	}
 
 	@PutMapping(value = "/{id}",produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
-	public ResponseEntity<DijagnozaDTO> updateDijagnoza(@RequestBody Dijagnoza d, @PathVariable("id") Long id) throws NotFoundException {
-		Dijagnoza d2 = service.update(id, d);
+	public ResponseEntity<DijagnozaDTO> updateDijagnoza(@RequestBody DijagnozaDTO d, @PathVariable("id") Long id, HttpServletRequest request) throws NotFoundException {
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra)){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
+		Dijagnoza d2 = service.update(id, new Dijagnoza(d));
 		return new ResponseEntity<DijagnozaDTO>(new DijagnozaDTO(d2), HttpStatus.OK);
 	}
 	
 	@DeleteMapping(value = "/{id}")
-	public ResponseEntity<DijagnozaDTO> deleteDijagnoza(@PathVariable("id") Long id){
+	public ResponseEntity<DijagnozaDTO> deleteDijagnoza(@PathVariable("id") Long id, HttpServletRequest request){
+		if (!(request.getSession().getAttribute("user") instanceof AdministratorKlinickogCentra)){
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+		}
 		service.delete(id);
 		return new ResponseEntity<DijagnozaDTO>(HttpStatus.NO_CONTENT);
 	}
